+++ bindings/python/CHANGELOG.md
@@ -66,6 +66,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `Wallet::{generate_mnemonic(), verify_mnemonic()}` since they're available from `Utils`;
 - `HD_WALLET_TYPE`, `HARDEN_MASK` constants;
+- `find_outputs()` method;
 
 ### Fixed